Features Changes: Default theme changed; Formalized the scope packaging for a future OSGi coding way; Changed the Coffee Resource Discovery, now fully Servlet Complaint; Fully compliant with Tomcat 6, JBoss 4.2.3 and JBoss 6; JavaScript DOM API Improvement. Bugfixes: Default Theme is now working on IE7+.

+++ coffee-sample-sugar/src/main/java/sugar/sample/Branch.java
@@ -6,25 +6,25 @@ import coffee.sugar.helper.ITreeItem;
 
 public class Branch implements ITreeItem {
 
-	private Integer id;
-	private Integer parentId;
+	private Long id;
+	private Long parentId;
 	private String label;
 	private Collection<? extends ITreeItem> children;
 	private Boolean hasChildren;
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
-	public void setParentId(Integer parentId) {
+	public void setParentId(Long parentId) {
 		this.parentId = parentId;
 	}
 
-	public Integer getParentId() {
+	public Long getParentId() {
 		return parentId;
 	}
